# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LipaBnb API",
        "description": "API documentation for LipaBnb application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get list of properties",
                "operationId": "e83a7fc5b1d98569892c10b7b0223879",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title or location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Pending",
                                "Active",
                                "Inactive"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Create a new property",
                "operationId": "70bf627c7a3b7593179b22a21956b41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "price",
                                    "bedrooms",
                                    "bathrooms",
                                    "guests",
                                    "property_type",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Luxury Beachfront Villa"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Malibu, California"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299
                                    },
                                    "bedrooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "bathrooms": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "guests": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "property_type": {
                                        "type": "string",
                                        "enum": [
                                            "apartment",
                                            "house",
                                            "villa",
                                            "studio",
                                            "condo"
                                        ],
                                        "example": "villa"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "Active",
                                            "Inactive"
                                        ],
                                        "example": "Active"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Beautiful beachfront villa with stunning ocean views"
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "WiFi",
                                            "Pool",
                                            "Parking"
                                        ]
                                    },
                                    "image": {
                                        "description": "Property image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Property"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Property": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Luxury Beachfront Villa"
                    },
                    "location": {
                        "type": "string",
                        "example": "Malibu, California"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 299
                    },
                    "bedrooms": {
                        "type": "integer",
                        "example": 3
                    },
                    "bathrooms": {
                        "type": "integer",
                        "example": 2
                    },
                    "guests": {
                        "type": "integer",
                        "example": 6
                    },
                    "property_type": {
                        "type": "string",
                        "enum": [
                            "apartment",
                            "house",
                            "villa",
                            "studio",
                            "condo"
                        ],
                        "example": "villa"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Pending",
                            "Active",
                            "Inactive"
                        ],
                        "example": "Active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Beautiful beachfront villa with stunning ocean views"
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "WiFi",
                            "Pool",
                            "Parking"
                        ]
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost:8000/storage/properties/image.jpg"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}